rename: 컴포넌트 폴더 소문자로 변경, 함수 이름 수정
