+++ src/components/IntroduceSections/Section/Interview/Principal/index.tsx
@@ -1,3 +1,4 @@
+import React from "react";
 import {
   Page__container,
   PrincipalPage,
@@ -9,15 +10,11 @@ import {
   Interview__wrapper,
   Interview__content,
   Principal__interview,
-} from "./Principal.style";
+} from "./principal.style";
 import { useInView } from "react-intersection-observer";
-import {
-  P_CONTENT_URL,
-  P_INFO_URL,
-  P_PERSON_URL,
-} from "components/IntroduceSections/assets";
+import { P_CONTENT_URL, P_INFO_URL, P_PERSON_URL } from "../../../assets";
 
-export default function Pricipal() {
+const Principal = () => {
   const [curPage, inView] = useInView({ threshold: 0.1 });
 
   return (
@@ -48,4 +45,6 @@ export default function Pricipal() {
       </Page__container>
     </PrincipalPage>
   );
-}
+};
+
+export default Principal;
