+++ src/components/IntroduceSections/assets/index.ts
@@ -1,9 +1,9 @@
-export { default as P_CONTENT_URL } from "./Principal/principal_content.png";
-export { default as P_INFO_URL } from "./Principal/principal_info.png";
-export { default as P_PERSON_URL } from "./Principal/principal.svg";
+export { default as P_CONTENT_URL } from "./principal/principal_content.png";
+export { default as P_INFO_URL } from "./principal/principal_info.png";
+export { default as P_PERSON_URL } from "./principal/principal.svg";
 
-export { default as S_CONTENT_URL } from "./Student/student_content.png";
-export { default as S_INFO_URL } from "./Student/student_info.png";
-export { default as S_PERSON_URL } from "./Student/student.svg";
+export { default as S_CONTENT_URL } from "./student/student_content.png";
+export { default as S_INFO_URL } from "./student/student_info.png";
+export { default as S_PERSON_URL } from "./student/student.svg";
 
-export { default as VICTIM_URL } from "./Result/victim.jpeg";
+export { default as VICTIM_URL } from "./result/victim.jpeg";
